move e2e tests inline
