+++ pkg/oauthflow/e2e_test.go
@@ -22,10 +22,7 @@ import (
 	"os"
 	"testing"
 
-	"github.com/agiledragon/gomonkey"
 	"github.com/go-rod/rod"
-	"github.com/sigstore/sigstore/pkg/oauthflow"
-	"github.com/skratchdot/open-golang/open"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 )
@@ -37,12 +34,12 @@ type OAuthSuite struct {
 func (suite *OAuthSuite) TestOauthFlow() {
 	urlCh := make(chan string)
 
-	// monkey patch this to not actually open a browser
-	patches := gomonkey.ApplyFunc(open.Run, func(input string) error {
+	oldOpener := browserOpener
+	browserOpener = func(input string) error {
 		urlCh <- input
 		return nil
-	})
-	defer func() { patches.Reset() }()
+	}
+	defer func() { browserOpener = oldOpener }()
 
 	go func() {
 		authCodeURL := <-urlCh
@@ -50,11 +47,11 @@ func (suite *OAuthSuite) TestOauthFlow() {
 		page.MustElement("body > div.dex-container > div > div > div:nth-child(2) > a > button").MustClick()
 	}()
 
-	idToken, err := oauthflow.OIDConnect(
+	idToken, err := OIDConnect(
 		os.Getenv("OIDC_ISSUER"),
 		os.Getenv("OIDC_ID"),
 		"",
-		oauthflow.DefaultIDTokenGetter,
+		DefaultIDTokenGetter,
 	)
 
 	require.Nil(suite.T(), err)
