+++ test/e2e/e2e-test.sh
@@ -62,6 +62,6 @@ export AWS_TLS_INSECURE_SKIP_VERIFY=1
 export OIDC_ISSUER=http://127.0.0.1:5556/auth
 export OIDC_ID=sigstore
 
-go test -tags e2e -count=1 ./...
+go test -tags e2e -count=1 ../../...
 
 cleanup
